backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
psycopg[binary]==3.1.17
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator
